requirements.txt
fastapi~=0.110.0
peewee
passlib[bcrypt]
pydantic[email]
python-jose